commonjs-ify NITC export
