+++ pages/nitcexport.html
@@ -2,9 +2,7 @@
 <html>
 <head>
 <meta charset="utf-8">
-<title>NITC  Export</title>
-<link rel="stylesheet" href="vendors/bootstrap/css/bootstrap.min.css">
-<link rel="stylesheet" href="resources/css/style.css" rel="stylesheet">
+<title>NITC Export</title>
 </head>
 <body class="small_foot_body">
 <div id="loading" style="visibility: hidden">
@@ -28,20 +26,17 @@
         </label>
       </div>
     </div>
-    <br>
+    <br />
       <input type="button" id="goButton" class="btn btn-primary btnSeccion" value="Download"></input>
   </form>
 </div>
 <footer class="small_foot">
   Designed &amp; developed by volunteers of the NSW SES.
-  Lighthouse is distributed under an <a href= "https://github.com/OSPFNeighbour/BeaconEnhanced/blob/master/LICENSE.md" rel="license">MIT Licence</a>.
+  Lighthouse is distributed under an
+  <a href= "https://github.com/OSPFNeighbour/BeaconEnhanced/blob/master/LICENSE.md"
+     rel="license">MIT Licence</a>.
 </footer>
-<script src="vendors/jquery/jquery-2.1.4.min.js" type="text/javascript"></script>
-<script src="vendors/bootstrap/js/bootstrap.min.js" type="text/javascript"></script>
-<script src="resources/js/nitcexport.js" type="text/javascript"></script>
-<script src="resources/js/shared_nitc_code.js" type="text/javascript"></script>
-<script src="resources/js/shared_json_code.js" type="text/javascript"></script>
 <script src="vendors/greensock-js/src/minified/TweenMax.min.js"></script>
-<script src="vendors/ElasticProgress/dist/elastic-progress.js"></script> 
+<script src="scripts/nitcexport.js" type="text/javascript"></script>
 </body>
 </html>
